+++ src/app/components/forms/input/date-picker/terra-date-picker.component.ts
@@ -12,7 +12,8 @@ import {
 import {
     IMyDateModel,
     IMyOptions,
-    MyDatePicker
+    MyDatePicker,
+    IMyInputFocusBlur
 } from 'mydatepicker';
 import { isNullOrUndefined } from 'util';
 import moment = require('moment');
@@ -151,22 +152,44 @@ export class TerraDatePickerComponent implements OnChanges, ControlValueAccessor
         if(!isNullOrUndefined(value) && typeof (value) === 'object')
         {
             this._value = value;
-
-            this.onTouchedCallback();
-            this.onChangeCallback(moment(value.jsdate).format());
         }
         else
         {
             this._value = null;
-
-            this.onTouchedCallback();
-            this.onChangeCallback(null);
         }
     }
 
     public clearDate():void
     {
         this.viewChildMyDatePicker.clearDate();
+    }
+
+    /**
+     * Is triggered on `ngModelChange` and executes `onChangeCallBack`
+     * @param value
+     */
+    protected onChange(value:IMyDateModel):void
+    {
+        if(!isNullOrUndefined(value))
+        {
+            this.onChangeCallback(moment(value.jsdate).format());
+        }
+        else
+        {
+            this.onChangeCallback(null);
+        }
+    }
+
+    /**
+     * Is triggered on `inputFocusBlur` and executes `onTouchedCallback` if a blur event is emitted
+     * @param event
+     */
+    protected onFocusOrBlur(event:IMyInputFocusBlur):void
+    {
+        if(event.reason === 2) // blur
+        {
+            this.onTouchedCallback();
+        }
     }
 
     private updateDatePickerOptions():void